New Development: No             JIRA Issue: No         Ready for Test: Yes      Interface Changes: Yes/No What Interface Changed: Please list interface changes          Test Programs: List test programs   Put in Release Notes: Yes/No              Module(s): Module Names change impacts.            Description: eliminate compiler warnings
git-svn-id: https://svn.cv.nrao.edu/svn/casa/active@14909 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ casacore/images/Images/ImageMetaData.cc
@@ -82,7 +82,7 @@ namespace casa { //# NAMESPACE CASA - BEGIN
 
     Int ImageMetaData::stokesPixelNumber(const String& stokesString) const {
     	Int pixNum = itsCoordinates.stokesPixelNumber(stokesString);
-    	if (pixNum >= nStokes()) {
+    	if (pixNum >= (Int)nStokes()) {
     		pixNum = -1;
     	}
     	return pixNum;
